requirements.txt
bottle==0.12.19
bottle-sqlite==0.2.0
Paste==3.5.0
PyJWT==2.3.0
six==1.16.0
